feat: using podman-desktop modals instead of custom one

+++ packages/backend/src/studio-api-impl.ts
@@ -115,8 +115,18 @@ export class StudioApiImpl implements StudioAPI {
     return this.catalogManager.getCatalog();
   }
 
-  async deleteLocalModel(modelId: string): Promise<void> {
-    await this.modelsManager.deleteLocalModel(modelId);
+  async requestRemoveLocalModel(modelId: string): Promise<void> {
+    const modelInfo = this.modelsManager.getLocalModelInfo(modelId);
+
+    podmanDesktopApi.window.showWarningMessage(
+      `Are you sure you want to delete ${modelId} ? The following files will be removed from disk "${modelInfo.file}".`,
+      'Confirm',
+      'Cancel'
+    ).then((result: string) => {
+      if(result === 'Confirm') {
+        this.modelsManager.deleteLocalModel(modelId);
+      }
+    });
   }
 
   async getModelsDirectory(): Promise<string> {
